requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
PyJWT==2.8.0
natsort==8.4.0
